feat: 관리자용 매장 목록 조회 API 구현

+++ src/main/java/com/verby/indp/domain/playlist/dto/response/FindStorePlaylistResponse.java
@@ -10,7 +10,7 @@ public record FindStorePlaylistResponse(
 
     public record CurrentSongItem(
         Long playlistSongId,
-        int playOrder,
+        double playOrder,
         String title,
         String artist,
         String vid,
@@ -31,7 +31,7 @@ public record FindStorePlaylistResponse(
 
     public record SongItem(
         Long playlistSongId,
-        int playOrder,
+        double playOrder,
         String title,
         String artist,
         Integer playTime,
